==> R/ENM/11b_area_sp_BR.R <==
# Credits ---------------------------

# Created by
# Luara Tourinho (https://github.com/luaratourinho)

# Last update: 20 Jul 2022


# Required packages

library(tidyverse)
library(dplyr)
library(raster)
library(rgdal)
library(rgeos)

# Read your species list
sp_names <- read.csv("./ENM/data/03_n_thinned_records.csv", 
                     stringsAsFactors = FALSE)
sp_names$species <-
  gsub(x = sp_names$species,
       pattern = " ",
       replacement = "_")

# Only charismatics with less than 15 occurrences

sp_names_l15  <- sp_names %>%
  subset(n_thin_10 < 15) %>%
  pull(species)

sp_names_h15  <- sp_names %>%
  subset(n_thin_10 >= 15) %>%
  pull(species)

sp_Es  <- read.csv("./tables/eco_services_binary.csv", stringsAsFactors = FALSE)

sp_Es$genus_epithet_IUCN <-
  gsub(x = sp_Es$genus_epithet_IUCN,
       pattern = " ",
       replacement = "_")

sp_Es_sub  <- sp_Es %>%
  subset(CHARISMATIC_AND_OR_CULTURAL_IMPORTANT_SPECIES == 1)%>%
  pull(genus_epithet_IUCN)

less15_charism  <- sp_names_l15 %>%
  subset(sp_names_l15 %in% sp_Es_sub)

# Species to run

species_to_run <- c(sp_names_h15, less15_charism)
species_to_run_ord <- sort(species_to_run)
sp_names <- species_to_run_ord[-c(204, 205)]

# sp_names <- unique(sp_names$species)
# sp_names <- sp_names[-c(248, 249)] #"Molossus_coibensis" (ENM files do not exist)

n <-length(sp_names)

BR <- shapefile("./Shp/Brasil_wgs84_diss.shp")

# Building table by a loop ---------------------------------------------------

Results_diff_area <- matrix(nrow = n, ncol = 19)
colnames(Results_diff_area) <- c("Species", "unchanged_percent_bin","loss_percent_bin", 
                                 "gain_percent_bin", "gain_loss_perc_pixel", 
                                 "unchanged", "loss", "gain", "total_cells",
                                 "ocorrencia_area_cu", "ocorrencia_area_fu",
                                 "loss_area3", "gain_area3", "unchaged_area3",
                                 "perc_loss_rel_pres", "perc_gain_rel_pres",
                                 "net_value_area", "net_value_perc",
                                 "percentage_change")

#"mean_diff_fut_cur_percent", "t", "p"


for(i in 1:n){
  
  # Read your current and future rasters
  # Binary and continuous projections
  cu <- raster(paste0("./ENM/outputs/", sp_names[i], "/results/", 
                      "CUR.bin_", sp_names[i], '.asc'))
  cu <- crop(cu, BR)
  cu <- mask(cu, BR)
  
  fu <- raster(paste0("./ENM/outputs/", sp_names[i], "/results/", 
                      '/Fut_bcc.bin_', sp_names[i], '.asc'))
  
  fu <- crop(fu, BR)
  fu <- mask(fu, BR)
  
  cu_cont <- raster(paste0("./ENM/outputs/", sp_names[i], "/results/", 
                           "CUR.cont_", sp_names[i], ".asc"))
  
  cu_cont <- crop(cu_cont, BR)
  cu_cont <- mask(cu_cont, BR)
  
  fu_cont <- raster(paste0("./ENM/outputs/", sp_names[i], "/results/", 
                           '/Fut_bcc.cont_', sp_names[i], '.asc'))
  
  fu_cont <- crop(fu_cont, BR)
  fu_cont <- mask(fu_cont, BR)
  
  
  #BINARY
  
  cu2 = cu
  cu2[cu2 == 1] <- 2
  # fut-pres
  # 1-2 = -1 -> Here should be 1-1 = 0, i.e., the two have not changed, STABLE
  # 1-0 = 1 -> GAIN
  # 0-2 = -2 ->  Here should be 0-1 = -1, i.e., LOSS
  # 0-0 = 0 -> inadequate area in both scenarios, so the next step we turn into NA
  
  diff_bin <- fu - cu2
  diff_bin[diff_bin == 0] <- NA
  
  
  
  # Values and percentage related to the number of pixels
  
  unchanged <- ncell(which(diff_bin[] == -1))
  loss <- ncell(which(diff_bin[] == -2))
  gain <- ncell(which(diff_bin[] == 1))
  total_cells <- unchanged + loss + gain
  
  unchanged_percent_bin <- (unchanged/total_cells)*100
  loss_percent_bin <- (loss/total_cells)*100
  gain_percent_bin <- (gain/total_cells)*100
  gain_loss_perc_pixel <- gain_percent_bin - loss_percent_bin
  
  
  # Reclassing to -1, zero and 1 to save the raster
  # This is a raster that you could present in your final work
  
  diff_bin2 = diff_bin
  diff_bin2[diff_bin2 == -1] <- 0
  diff_bin2[diff_bin2 == -2] <- -1
  
  writeRaster(diff_bin2, filename = paste0("./ENM/Diff_raster/", sp_names[i],
                                           "_diff_bin.tif"),
              format="GTiff", overwrite=T)
  
  
  # CONTINUOUS
  
  # Evaluating the difference between future and current continuous projections
  # Mean of difference - Negative value means loss, positive value means gain
  
  # testt <- t.test(fu_cont[], cu_cont[], paired=T)
  # chars <- capture.output(print(testt))
  # mean_diff_estimate <- as.data.frame(testt$estimate)
  # mean_diff_fut_cur <- mean_diff_estimate[1,]
  # mean_diff_fut_cur_percent <- mean_diff_fut_cur*100
  # mean_diff_conf.int <- as.data.frame(testt$conf.int)
  # t <- testt$statistic
  # p <- testt$p.value
  
  diff_cont <- fu_cont - cu_cont
  
  writeRaster(diff_cont, filename = paste0("./ENM/Diff_raster/", sp_names[i],
                                           "_diff_cont.tif"),
              format="GTiff", overwrite=T)
  
  
  
  # AREA --------------------------------------------------------------------
  
  # To calculate the areas, we normally use binary results
  
  # Values and percentage related to the area in km2
  
  #current
  cu_area = cu
  cu_area[cu_area == 0] <- NA
  r = raster(nrow = cu_area@nrows, ncol = cu_area@ncols, xmn = cu_area@extent@xmin, 
             xmx = cu_area@extent@xmax, ymn = cu_area@extent@ymin, ymx = cu_area@extent@ymax) 
  # calculate area of each cell (the area changes along the latitudes / longitudes)
  x = raster::area(r) 
  #plot(x)
  ocorrencia_area = x * cu_area
  ocorrencia_area_cu <- cellStats(ocorrencia_area, stat='sum', na.rm=TRUE, asSample=TRUE)
  
  #future
  fu_area = fu 
  fu_area[fu_area == 0] <- NA
  ocorrencia_area2 = x * fu_area
  ocorrencia_area_fu <- cellStats(ocorrencia_area2, stat='sum', na.rm=TRUE, asSample=TRUE)
  
  # Expected area to contract (loss)
  
  loss_area = diff_bin 
  loss_area[loss_area == -1] <- NA
  loss_area[loss_area == 1] <- NA
  loss_area[loss_area == -2] <- 1
  loss_area2 = x * loss_area
  loss_area3 <- cellStats(loss_area2, stat='sum', na.rm=TRUE, asSample=TRUE)
  # Percentage of loss related to the current area
  perc_loss_rel_pres <- (loss_area3*100)/ocorrencia_area_cu
  
  # Expected area to expand (gain)
  
  gain_area = diff_bin 
  gain_area[gain_area == -1] <- NA
  gain_area[gain_area == -2] <- NA
  gain_area2 = x * gain_area
  gain_area3 <- cellStats(gain_area2, stat='sum', na.rm=TRUE, asSample=TRUE)
  # Percentage of gain related to the current area
  perc_gain_rel_pres <- (gain_area3*100)/ocorrencia_area_cu
  
  
  # Expected area not to change (unchanged)
  
  unchaged_area = diff_bin 
  unchaged_area[unchaged_area == 1] <- NA
  unchaged_area[unchaged_area == -2] <- NA
  unchaged_area[unchaged_area == -1] <- 1
  unchaged_area2 = x * unchaged_area
  unchaged_area3 <- cellStats(unchaged_area2, stat='sum', na.rm=TRUE, asSample=TRUE)
  
  
  # How much did change?
  
  # Net value in area and in percentage
  net_value_area <- gain_area3 - loss_area3
  net_value_perc <- perc_gain_rel_pres - perc_loss_rel_pres
  
  # Percentage of Change = (FV/IV - 1) * 100
  # Where FV is the final value of the transaction, whereas IV refers to the initial value.
  # The increase was ((700/500) - 1) * 100 = 0,4 × 100 = 40%.
  # The loss was ((500/700) - 1)* 100 = -29%
  percentage_change <- ((ocorrencia_area_fu/ocorrencia_area_cu) - 1)*100
  
  
  
  Results_diff_area[i, ] <- c(sp_names[i], unchanged_percent_bin, loss_percent_bin,
                              gain_percent_bin, gain_loss_perc_pixel, 
                              unchanged, loss, gain, total_cells,ocorrencia_area_cu, ocorrencia_area_fu,
                              loss_area3, gain_area3, unchaged_area3,
                              perc_loss_rel_pres, perc_gain_rel_pres,
                              net_value_area, net_value_perc,
                              percentage_change)
  
  #mean_diff_fut_cur_percent, t, p
  
}

head(Results_diff_area)

write.csv(Results_diff_area, "./ENM/11_area_sp_BR_less15onlycharis.csv", row.names = F)
